requirements.txt
astroid==2.8.0
autopep8==1.5.7
bcrypt==3.2.0
black==21.9b0
blinker==1.4
cffi==1.14.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
flake8==3.9.2
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pylint==2.11.1
regex==2021.9.24
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
